+++ InventoryManagement.Data/Models/ExpenseTransportationList.cs
@@ -7,7 +7,7 @@
         public int NumberOfPerson { get; set; }
         public string ExpenseFor { get; set; }
         public string Vehicle { get; set; }
-        public double ExpenseAmount { get; set; }
+        public decimal ExpenseAmount { get; set; }
         public virtual ExpenseTransportation ExpenseTransportation { get; set; }
 
     }
