+++ InventoryManagement.Data/Models/Product.cs
@@ -19,7 +19,7 @@ namespace InventoryManagement.Data
         public string Description { get; set; }
         public string Warranty { get; set; }
         public string Note { get; set; }
-        public double SellingPrice { get; set; }
+        public decimal SellingPrice { get; set; }
         public DateTime InsertDate { get; set; }
         public virtual ProductCatalog ProductCatalog { get; set; }
         public virtual ICollection<ProductStock> ProductStock { get; set; }
