+++ InventoryManagement.Data/Models/Purchase.cs
@@ -15,12 +15,12 @@ namespace InventoryManagement.Data
         public int RegistrationId { get; set; }
         public int VendorId { get; set; }
         public int PurchaseSn { get; set; }
-        public double PurchaseTotalPrice { get; set; }
-        public double PurchaseDiscountAmount { get; set; }
-        public double PurchaseDiscountPercentage { get; set; }
-        public double PurchasePaidAmount { get; set; }
-        public double PurchaseReturnAmount { get; set; }
-        public double PurchaseDueAmount { get; set; }
+        public decimal PurchaseTotalPrice { get; set; }
+        public decimal PurchaseDiscountAmount { get; set; }
+        public decimal PurchaseDiscountPercentage { get; set; }
+        public decimal PurchasePaidAmount { get; set; }
+        public decimal PurchaseReturnAmount { get; set; }
+        public decimal PurchaseDueAmount { get; set; }
         public string PurchasePaymentStatus { get; set; }
         public string MemoNumber { get; set; }
