+++ InventoryManagement.Repository/Repositories/Customer/ICustomerRepository.cs
@@ -15,7 +15,7 @@ namespace InventoryManagement.Repository
         void CustomUpdate(CustomerAddUpdateViewModel model);
         void UpdatePaidDue(int id);
         Task<ICollection<CustomerListViewModel>> SearchAsync(string key);
-        double TotalDue();
+        decimal TotalDue();
         ICollection<CustomerDueViewModel> TopDue(int totalCustomer);
         DataResult<CustomerDueViewModel> TopDueDataTable(DataRequest request);
     }
