+++ InventoryManagement.Repository/Repositories/ProductStock/ProductStockRepository.cs
@@ -91,7 +91,7 @@ namespace InventoryManagement.Repository
             return Context.ProductStock.Include(s => s.Product).Where(s => codes.Contains(s.ProductCode)).ToListAsync();
         }
 
-        public double StockProductPurchaseValue()
+        public decimal StockProductPurchaseValue()
         {
             return Context.ProductStock.Include(s => s.PurchaseList)
                        .Where(s => !s.IsSold)?.Sum(s => s.PurchaseList.PurchasePrice) ?? 0;
